ferrokey product: core state machine, protocol, uinput backend, layouts, surface backends, custom Slint platform, ferrokeyd broker, ferrokey UI
- ferrokey-core: KeyboardState (latched/locked modifiers, double-tap-to-lock,
  rollover caps), deterministic RepeatEngine, Layout model (PhysicalKey/
  KeySymbol separation), KeyboardDriver + KeySink/TextSink
- ferrokey-protocol: FK01 binary codec with hostile-input-resistant streaming
  decoder, SO_PEERCRED identity
- ferrokey-uinput: explicit capability set, evdev device, emit helpers, held-key
  ledger (release-all recovery contract)
- ferrokey-layouts: us/us-intl/gb/de/fr/dvorak YAML data layouts, xkbcommon seam
- ferrokey-surface: capability detection, Wayland layer-shell (keyboard_interactivity
  = none), X11 WM_HINTS.input=False, custom Slint platform over both
- ferrokeyd: unprivileged-broker daemon, SO_PEERCRED whitelist, token-bucket rate
  limiting, hostile-input rejection, sigwait shutdown
- ferrokey: Slint UI wired to core driver over the daemon link, text-mode input,
  degraded-mode reporting

+++ crates/ferrokey-uinput/src/emit.rs
@@ -12,7 +12,7 @@ use std::io;
 /// One `EV_KEY` event. The caller batches and lets
 /// [`VirtualDevice::emit`] add the trailing `SYN_REPORT`.
 pub fn key_event(code: KeyCode, value: i32) -> InputEvent {
-    InputEvent::new(EventType::KEY.0 as u16, code.code() as u16, value)
+    InputEvent::new(EventType::KEY.0, code.code(), value)
 }
 
 /// A press event (`EV_KEY` value=1).
@@ -61,7 +61,7 @@ mod tests {
     fn event_construction() {
         let down = key_down_event(KeyCode::KEY_A);
         assert_eq!(down.event_type(), EventType::KEY);
-        assert_eq!(down.code(), u16::from(KeyCode::KEY_A.code()));
+        assert_eq!(down.code(), KeyCode::KEY_A.code());
         assert_eq!(down.value(), 1);
 
         let up = key_up_event(KeyCode::KEY_A);
